Add files via upload AlexNet CNN implementation on dacl10k dataset for training the model.
